Revert test changes from PR #1
We are no longer changing behavior for how PaymentRequest handles its details
constructor argument, so the test changes from PR #1 are no longer valid. Revert these.

There is a better way of changing PaymentRequestUpdateEvent, so revert all the
changes from PR #1 and then re-do the tests in a following commit.