Fix option checks to use boolValue
Every place in this file where option values are checked is incorrect. All
that's being tested is that the option exists. If the options being tested exist
but are false, the current code treats them as true.

In order to test whether these options are true or false, we have to use
boolValue.